Fixes to Arduino template
Test by loading all into Arduino IDE and compiling

+++ test/sampleData/TCS3472.h
@@ -105,7 +105,7 @@ class TCS3472 {
          * Enable RGBC and Power
 
          */
-        uint8_t setinit(uint8_t data);
+        int setinit(uint8_t data);
